% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelling.R
\name{apply_relabelling}
\alias{apply_relabelling}
\alias{apply_relabelling.bmoe_fit}
\alias{table_permutations}
\title{Apply Relabelling Scheme to \code{\link{bmoe}}}
\usage{
apply_relabelling(object, permutations = NULL)

\method{apply_relabelling}{bmoe_fit}(object, permutations = NULL)

table_permutations(permutations)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{permutations}{list. See \strong{Permutations} section.}
}
\value{
\code{\link{bmoe_fit}} object with switched labels.
}
\description{
Allows swapping of component labels, per MCMC iteration, to mitigate label
switching.
}
\section{Functions}{
\itemize{
\item \code{table_permutations()}: Calculate summary of a \code{permutations} object.

}}
\section{Permutations}{

Permutations are stored in a list where \eqn{i^{th}} elements corresponds to
the \eqn{i^{th}} chain.

Each element is a matrix with one permutation per row for each iteration.
Note that
\itemize{
\item A vector can be supplied to apply the same scheme to all MCMC output.
\item A matrix can be supplied to apply the same relabelling to all chains.
\item \code{NULL} can be supplied to use \code{\link[=boys_henderson_2002]{boys_henderson_2002()}}.
}

A permutation of \code{c(3, 1, 2)} denotes relabelling of
\itemize{
\item \code{1 -> 3}
\item \code{2 -> 1}
\item \code{3 -> 2}
}
}

