% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{artificial_Surv}
\alias{artificial_Surv}
\title{Artificial Censoring}
\usage{
artificial_Surv(x, q_cens)
}
\arguments{
\item{x}{vector. Uncensored numerical data.}

\item{q_cens}{numeric [0, 1]. Optional artificial left-censoring level.}
}
\value{
Surv object.
}
\description{
Convert numeric vector to a \code{survival::Surv} object by censoring at
chosen quantiles.
}
\keyword{internal}
