% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{bmoe-allocation-vis}
\alias{bmoe-allocation-vis}
\alias{visualise_allocations_with_response_data}
\alias{visualise_allocation_data}
\alias{visualise_response_data}
\title{Visualise Posterior Group Membership, Ordered by Rank}
\usage{
visualise_allocations_with_response_data(object, .chain, .rank_by)

visualise_allocation_data(object, .chain)

visualise_response_data(object, .plot_by, .rank_by)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{.chain}{integer. Which chain indexes to use as posterior allocations.}

\item{.rank_by}{expression. Column to use in ordering (y-axis).}

\item{.plot_by}{expression. Column to be plotted (x-axis).}
}
\description{
Combination of allocation probabilities and quantile plots
}
\details{
For consistency between each plot, ordering is based on a single chosen
response variable.
}
\note{
Both \code{.plot_by} and \code{.rank_by} can be censored (\verb{<Surv>}).
}
\section{Posterior Allocation Plot}{

Histogram showing the corresponding proportions of each component
allocation per observation.
}

\section{NA}{

Histogram showing the corresponding proportions of each component
allocation per observation.
}

