% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-args.R
\name{bmoe-args}
\alias{bmoe-args}
\alias{bmoe_prior}
\alias{bmoe_jags_n}
\alias{bmoe_inits}
\title{Bayesian MoE (\code{bmoe}) Arguments}
\usage{
bmoe_prior(k, ...)

bmoe_jags_n(...)

bmoe_inits(seed_base, .RNG.name = "base::Wichmann-Hill", ...)
}
\arguments{
\item{k}{integer. Assumed number of components.}

\item{...}{Overwrite default values.}

\item{seed_base}{numeric. Seed are set equal to \code{seed_base} multiplied by
chain index.}

\item{.RNG.name}{character. JAGS RNG method.}
}
\description{
Helper functions to assist in supplying arguments to \code{\link[=bmoe]{bmoe()}}.
}
\section{Prior}{

Hyper-parameters that must be passed to the \code{prior} argument are in bold.
\itemize{
\item \code{k}, number of components, assumed known.
\item \code{regr} is element-wise IID Normal with 0 mean and \strong{regr_prec} precision.
\item \code{wt} is element-wise IID Normal with 0 mean and \strong{wt_prec} precision.
\item \code{prec} is element-wise IID Gamma with \strong{prec_shape} and \strong{prec_rate}.
}
}

\section{JAGS Controls}{

\itemize{
\item \code{n.adapt} controls number of discarded samples in adaptation stage.
\item \code{n.update} controls number of discarded samples in warm-up stage.
\item \code{n.iter} controls how many samples are saved.
\item \code{n.thin} controls thinning, where only every \eqn{n^th} sample is kept.
\item \code{n.chains} controls number of chains.
}
}

\section{JAGS Initial Values}{

Optional chain starting values, as described in \code{\link[rjags:jags.model]{rjags::jags.model()}}.
}

