% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{bmoe-deprec}
\alias{bmoe-deprec}
\alias{calculate_x_regr}
\title{Deprecated Functions}
\usage{
calculate_x_regr(object, new_data)
}
\description{
These functions should not be used and will be remove in \code{v1.0.0}.
}
