% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-allocated-quantities.R
\name{bmoe-extract-allocated}
\alias{bmoe-extract-allocated}
\alias{extract_allocated_y_posterior_mean}
\alias{extract_allocated_y_posterior_sd}
\title{Extract Allocated Quantities from Model Fit}
\usage{
extract_allocated_y_posterior_mean(object, new_data, z = NULL)

extract_allocated_y_posterior_sd(object, new_data, z = NULL)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}

\item{z}{\code{\link{bmoe_array}}.
MCMC samples used to integrate out components.

Defaults to \code{\link[=extract_allocation_samples]{extract_allocation_samples()}}.}
}
\value{
All quantities returned as \code{\link[=bmoe_array]{bmoe_array()}} of varying dimension.

Note that \code{n_s_new} denotes the number of rows in \code{new_data} or the
observed data when \code{new_data} is \code{NULL} valued.

\itemize{
\item \code{extract_allocated_y_posterior_mean()}
\itemize{
\item \code{dim(.) = c(n_iters, n_chains, n_s_new, n_y)}.
}
}

\itemize{
\item \code{extract_allocated_y_posterior_sd()}
\itemize{
\item Serves different purpose to \code{\link[=extract_y_posterior_sd]{extract_y_posterior_sd()}} since
parameters are recycled per observation in \code{new_data}.
\item \code{dim(.) = c(n_iters, n_chains, n_s_new, n_y)}.
}
}
}
\description{
Allocated versions of generated quantities where allocation samples, \code{z}.
}
