% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihood.R
\name{bmoe-log-lik}
\alias{bmoe-log-lik}
\alias{extract_pointwise_log_lik}
\alias{extract_log_lik}
\title{Extract Pointwise Log Likelihood}
\usage{
extract_pointwise_log_lik(object, new_data)

extract_log_lik(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the trained model (special case).}
}
\description{
Offline computation of log likelihood for each response variable.
}
\seealso{
The \code{loo} package for using these quantities in model comparison.
}
