% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-package.R
\docType{package}
\name{bmoe-package}
\alias{bmoe-package}
\title{Bayesian Mixture of Experts Package}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}
}
\description{
Bayesian Mixture of Experts (MoE) applied to possibly left-censored
linear regression.
}
\section{Model Specification}{

TODO
}

\seealso{
Useful links:
\itemize{
  \item \url{https://ncljoshcowley.github.io/bmoe/}
}

}
\author{
\strong{Maintainer}: Josh Cowley \email{j.cowley1@ncl.ac.uk}

}
