% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{bmoe-plot}
\alias{bmoe-plot}
\alias{autoplot.bmoe_fit}
\title{List of MoE Analysis Plot Objects}
\usage{
\method{autoplot}{bmoe_fit}(object, varname, type = "none", ..., new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{varname}{character. Variable name defined in the model.}

\item{type}{choice. Name of the supported plot type including
\itemize{
\item \code{"none"}, plots are to be empty until the user adds layers.
\item \code{"density"}, useful distribution visualisation.
\item \code{"trace"}, useful for verifying chain convergence.
\item \code{"acf"}, autocorrelation of the MCMC samples at discrete lags.
Alternatively one can pass a \strong{list} of custom layers to be added.
}}

\item{...}{Extra arguments silently ignored.}

\item{new_data}{Passed to \code{\link[=extract_log_lik]{extract_log_lik()}}; ignored otherwise.}
}
\description{
Creates a nested list of \code{ggplot2} objects with partitioned data, requires
the user to add \link[=mcmc_layer]{layers}.
}
