% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe.R
\name{bmoe}
\alias{bmoe}
\alias{bmoe_fit}
\alias{bmoe.bmoe_sim}
\alias{bmoe.formula}
\title{Bayesian Mixture of Experts Linear Regression.}
\usage{
bmoe(object, ..., prior, jags_n = NULL, inits = NULL)

\method{bmoe}{bmoe_sim}(object, ..., prior, jags_n = NULL, inits = NULL)

\method{bmoe}{formula}(object, data, ..., prior, jags_n = NULL, inits = NULL)
}
\arguments{
\item{object}{object. A formula or simulated object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{prior}{named list. See \strong{Prior} section.}

\item{jags_n}{named list. See \strong{JAGS Controls} section.}

\item{inits}{list. Passed to \code{rjags::jags.model()}.}

\item{data}{data frame. To be used in modelling.}
}
\description{
Fits the model in JAGS.
}
\section{Prior}{

Hyper-parameters that must be passed to the \code{prior} argument are in bold.
\itemize{
\item \code{k}, number of components, assumed known.
\item \code{regr} is element-wise IID Normal with 0 mean and \strong{regr_prec} precision.
\item \code{wt} is element-wise IID Normal with 0 mean and \strong{wt_prec} precision.
\item \code{prec} is element-wise IID Gamma with \strong{prec_shape} and \strong{prec_rate}.
}
}

\section{JAGS Controls}{

\itemize{
\item \code{n.adapt} controls number of discarded samples in adaptation stage.
\item \code{n.update} controls number of discarded samples in warm-up stage.
\item \code{n.iter} controls how many samples are saved.
\item \code{n.thin} controls thinning, where only every \eqn{n^th} sample is kept.
\item \code{n.chains} controls number of chains.
}
}

