% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-predict.R
\name{bmoe_array_to_pred_tb}
\alias{bmoe_array_to_pred_tb}
\title{Convert \code{bmoe_array} to Prediction Table}
\usage{
bmoe_array_to_pred_tb(x, .key)
}
\arguments{
\item{x}{\code{\link{bmoe_array}} object.}

\item{.key}{character. Names to use in output, length should be \code{dim(x)[4]}.}
}
\description{
Creates a list column per response (or class) where each element is a draws
tibble containing \code{.iter}, \code{.chain} and \code{.value} for a single parameter.
}
\keyword{internal}
