% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-yx.R
\name{bmoe_yx}
\alias{bmoe_yx}
\title{Mixture of Experts Implementation}
\usage{
bmoe_yx(y_list, x_regr, x_wt, prior, jags_n, inits = NULL)
}
\arguments{
\item{y_list}{list. Response data per element; length is 1 when univariate.}

\item{x_regr, x_wt}{matrix. Regression and weighting matrices.}

\item{prior}{named list. See \strong{Prior} section.}

\item{jags_n}{named list. See \strong{JAGS Controls} section.}

\item{inits}{list. Passed to \code{rjags::jags.model()}.}
}
\description{
Implementation of Mixture of Experts models.
}
\keyword{internal}
