% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelling.R
\name{boys_henderson_2002}
\alias{boys_henderson_2002}
\title{Relabelling Algorithm, Boys & Henderson (2002)}
\usage{
boys_henderson_2002(z, start = 2, z_max = max(z))
}
\arguments{
\item{z}{\code{bmoe_array}. Component allocation output from \code{\link{bmoe_fit}}.}

\item{start}{integer. Required starting point for this algorithm.}

\item{z_max}{integer. Maximum possible value of \code{z}.}
}
\value{
See \strong{Permutations} section in \code{\link[=apply_relabelling]{apply_relabelling()}}.
}
\description{
Relabelling algorithm detailed in paper "On determining the order of Markov
dependence of an observed process" (Boys & Henderson, 2002).
}
\details{
Read paper for full details; aim is to minimise 'disagreement',

\eqn{
   D
       = - \sum_{i = 1}^n \mathbb{I} \left(
           \nu_m(z_i^{(m)}) = \hat{z}_i^{(m - 1)}
       \right)
}

where
\itemize{
\item \eqn{\nu_m} is a candidate permutation to be chosen;
\item \eqn{\hat{z}} is the marginal posterior mode of all MCMC iterations
up to that point.
\item relabelling is applied per MCMC iterations, \eqn{m = m_0, \dots, M} where
the user defines \eqn{m_0} using the \code{start} argument.
}
}
\references{
Boys, R. J. & Henderson, D. A. 2002
On determining the order of markov dependence of an observed process
governed by a hidden markov model.
Scientific Programming 10 (3), 241–251.
}
