% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_component_samples}
\alias{calculate_component_samples}
\title{Simulate New Component Allocation Samples}
\usage{
calculate_component_samples(wt, x_wt, varname)
}
\arguments{
\item{wt}{\code{\link{bmoe_array}}. Model contribution to probabilities}

\item{x_wt}{matrix. Data contribution to probabilities}

\item{varname}{character. Variable name of the stored quantity.}
}
\description{
Simulate New Component Allocation Samples
}
\seealso{
Other Posterior calculation functions: 
\code{\link{calculate_posterior_params}()}
}
\concept{Posterior calculation functions}
