% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_posterior_params}
\alias{calculate_posterior_params}
\alias{calculate_posterior_means}
\alias{calculate_posterior_sds}
\title{Calculate Model Fit's Posterior Mean and Standard Deviation}
\usage{
calculate_posterior_params(object, new_data, separate = FALSE)

calculate_posterior_means(z, regr, x_regr, varname)

calculate_posterior_sds(z, prec, varname)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame. Data used to create a design matrices.
Setting \code{new_data = NULL} utilises the trained model (special case).}

\item{separate}{logical. Splits the output into multiple arrays when \code{TRUE}.}

\item{z}{\code{\link{bmoe_array}}. Component allocation (latent variable) output.}

\item{regr}{\code{\link{bmoe_array}}. Regression coefficient output.}

\item{x_regr}{matrix. Required design matrix.}

\item{varname}{character. Variable name of the stored quantity.}

\item{prec}{\code{\link{bmoe_array}}. Precision matrix output.}
}
\description{
Calculate Model Fit's Posterior Mean and Standard Deviation

Calculate Model Fit's Posterior Mean(s)

Calculate Model Fit's Posterior Standard Deviation(s)
}
\seealso{
Other Posterior calculation functions: 
\code{\link{calculate_component_samples}()}
}
\concept{Posterior calculation functions}
