% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-parse-formula.R
\name{create_summand_from_symbols}
\alias{create_summand_from_symbols}
\title{Create Call From (List of) Symbols}
\usage{
create_summand_from_symbols(x)
}
\arguments{
\item{x}{list. Expressions to be added together.}
}
\description{
Output will be \code{expression(..1 + ..2)} based on input \code{list(..1, ..2)}.
}
\keyword{internal}
