% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-likelihood.R
\name{d_lcens_norm}
\alias{d_lcens_norm}
\title{Likelihood of Left-Censored Normally Distributed Vector}
\usage{
d_lcens_norm(x, mean, sd)
}
\arguments{
\item{x}{object. Expected to be left-censored \code{Surv}.}

\item{mean, sd}{numeric. Passed to \code{stats::dnorm} or \code{stats::pnorm}.}
}
\description{
Likelihood of Left-Censored Normally Distributed Vector
}
\note{
Passing non-\code{Surv} objects lead to \code{stats::dnorm(...)}, not an error.
}
\keyword{internal}
