% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-log-lik.R
\name{.extract_pointwise_log_lik}
\alias{.extract_pointwise_log_lik}
\title{Alternate approach to \code{\link[=extract_pointwise_log_lik]{extract_pointwise_log_lik()}}}
\usage{
.extract_pointwise_log_lik(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}
}
\description{
Retained as this may be more stable, yet slower, than the exported version.
}
\keyword{internal}
