% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_bmoe_fit}
\alias{example_bmoe_fit}
\title{Example Bayesian Mixture of Experts Model Fit}
\usage{
example_bmoe_fit(...)
}
\arguments{
\item{...}{Either \code{prior}, \code{jags_n} \code{inits} or passed to \code{simulate_bmoe}.}
}
\description{
Example Bayesian Mixture of Experts Model Fit
}
\seealso{
\code{\link[=bmoe]{bmoe()}}.
}
