% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_draws}
\alias{extract_draws}
\alias{extract_draws.bmoe_fit}
\alias{extract_draws.bmoe_simstudy}
\title{Extract Posterior Samples from Bayesian Mixture of Experts}
\usage{
extract_draws(object, varname, ...)

\method{extract_draws}{bmoe_fit}(object, varname, ...)

\method{extract_draws}{bmoe_simstudy}(object, varname, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{varname}{character. Variable name defined in the model.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Extract Posterior Samples from Bayesian Mixture of Experts
}
\section{Simulation Study}{

For models fit to a \code{bmoe_sim} object, a \verb{$.truth} column may also be added.

This functionality is conditional on the assumed \eqn{K} in \code{prior} being
the same as the \eqn{K} used to simulate the data.
}

