% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihood.R
\name{extract_posterior_expectations}
\alias{extract_posterior_expectations}
\title{Extract Posterior Expectation(s)}
\usage{
extract_posterior_expectations(object, x_regr)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{x_regr}{matrix. Design matrix to use in the calculation.}
}
\value{
List of MCMC arrays denoting the mean vector, one element per
response variable.
}
\description{
Offline calculation of posterior means, denoted \code{mean} in JAGS file.
}
\details{
Note that while \code{x_regr} and \code{regr} lead to a mean matrix of \code{n_s} rows
and \code{K} columns, we flatten this structure into a \code{n_s} vector by using
the component allocation samples, \code{z}.
}
