% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-array.R
\name{extract_single_draw}
\alias{extract_single_draw}
\title{Extract Single Draw}
\usage{
extract_single_draw(x, iter, chain)
}
\arguments{
\item{x}{\code{\link{bmoe_array}} object.}

\item{iter, chain}{integer. MCMC indexes to extract from.}
}
\description{
Extract a single draw for specified iteration and chain.
}
\note{
Only the MCMC dimensions are \link[=drop]{dropped}.
}
\keyword{internal}
