% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-autoplot.R
\name{extract_term_indices}
\alias{extract_term_indices}
\title{Extract Term Indices}
\usage{
extract_term_indices(x, dnames = NULL)
}
\arguments{
\item{x}{character vector. Matches \code{foo[...]} where only \code{...} are used.}

\item{dnames}{character vector. Output column names.}
}
\description{
Convert term, say \code{regr[a, b, c]}, to df with column for each dimension.
}
\keyword{internal}
