% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-getters.R
\name{get_dims_from_bmoe_fit}
\alias{get_dims_from_bmoe_fit}
\title{Get Data and Parameter Dimensions}
\usage{
get_dims_from_bmoe_fit(object, id)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{id}{choice. One of
\itemize{
\item \code{n_iters}: Number of iterations.
\item \code{n_chains}: Number of chains.
\item \code{n_s}: Number of observations.
\item \code{n_y}: Number of response variables.
\item \code{n_k}: Assumed number of components.
\item \code{p_regr}: Number of regression predictors (including intercept).
\item \code{p_wt}: Number of weighting predictors (including intercept).
}}
}
\description{
Helper to access names from data (\code{y}, \code{x}, \code{k}) to be displayed to users.
}
\keyword{internal}
