% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-layer.R
\name{get_layer}
\alias{get_layer}
\alias{get_truth_layer}
\alias{layer_density}
\alias{layer_trace}
\alias{layer_acf}
\title{Get \code{ggplot2} Layer(s) for Plotting Functionality}
\usage{
get_layer(type)

get_truth_layer(type)

layer_density()

layer_trace()

layer_acf()
}
\arguments{
\item{type}{choice. Name of the supported plot type including
\itemize{
\item \code{"none"}, plots are to be empty until the user adds layers.
\item \code{"density"}, useful distribution visualisation.
\item \code{"trace"}, useful for verifying chain convergence.
\item \code{"acf"}, autocorrelation of the MCMC samples at discrete lags.
}}
}
\description{
Get \code{ggplot2} Layer(s) for Plotting Functionality
}
\keyword{internal}
