% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe.R
\name{get_names_from_bmoe_fit}
\alias{get_names_from_bmoe_fit}
\title{Get Human Readable Names}
\usage{
get_names_from_bmoe_fit(object)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}
}
\description{
Helper to access names from data (\code{y}, \code{x}, \code{k}) to be displayed to users.
}
