% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert.R
\name{get_names_from_mixexpert}
\alias{get_names_from_mixexpert}
\title{Get Human Readable Names}
\usage{
get_names_from_mixexpert(object)
}
\arguments{
\item{object}{Object that inherits from \code{\link{mixexpert}}.}
}
\description{
Helper to access names from data (\code{y}, \code{x}, \code{k}) to be displayed to users.
}
