% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe.R
\name{get_x_regr_from_bmoe_fit}
\alias{get_x_regr_from_bmoe_fit}
\title{Get Regression Design Matrix (\code{x_regr})}
\usage{
get_x_regr_from_bmoe_fit(object, data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}
}
\description{
Helper to access \code{x_regr} and verify a common RHS to regression formulas.
}
