% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-plots.R
\name{impl_c_alloc_vis}
\alias{impl_c_alloc_vis}
\title{Implementation of \code{component_allocation_visualisation}}
\usage{
impl_c_alloc_vis(x, labels)
}
\arguments{
\item{x}{integer-valued matrix. Rows (columns) denote iterations (indices).}

\item{labels}{character. Labels to use in plot and \verb{$data}.}
}
\description{
Create \code{ggplot2} object on a chain-by-chain basis to help with \code{gc()}.
}
\keyword{internal}
