% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{mcmc_layer}
\alias{mcmc_layer}
\alias{layer_density}
\alias{layer_trace}
\alias{layer_acf}
\title{Get \code{ggplot2} Layer(s) for Plotting MCMC Output.}
\usage{
mcmc_layer(type, .truth = FALSE)

layer_density(.truth = FALSE)

layer_trace(.truth)

layer_acf()
}
\arguments{
\item{type}{choice. Name of the supported plot type including
\itemize{
\item \code{"none"}, plots are to be empty until the user adds layers.
\item \code{"density"}, useful distribution visualisation.
\item \code{"trace"}, useful for verifying chain convergence.
\item \code{"acf"}, autocorrelation of the MCMC samples at discrete lags.
Alternatively one can pass a \strong{list} of custom layers to be added.
}}
}
\description{
Get \code{ggplot2} Layer(s) for Plotting MCMC Output.
}
\keyword{internal}
