% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-args.R
\name{mixexpert-args}
\alias{mixexpert-args}
\alias{mixexpert_prior}
\alias{mixexpert_jags_n}
\title{MoE (\code{mixexpert}) Arguments}
\usage{
mixexpert_prior(k, ...)

mixexpert_jags_n(...)
}
\arguments{
\item{k}{integer. Assumed number of components.}

\item{...}{Overwrite default values.}
}
\description{
Helper functions to assist in supplying arguments to \code{\link[=mixexpert]{mixexpert()}}.
}
\section{Prior}{

Hyper-parameters that must be passed to the \code{prior} argument are in bold.
\itemize{
\item \code{k}, number of components, assumed known.
\item \code{regr} is element-wise IID Normal with 0 mean and \strong{regr_prec} precision.
\item \code{wt} is element-wise IID Normal with 0 mean and \strong{wt_prec} precision.
\item \code{prec} is element-wise IID Gamma with \strong{prec_shape} and \strong{prec_rate}.
}
}

\section{JAGS Controls}{

\itemize{
\item \code{n.adapt} controls number of discarded samples in adaptation stage.
\item \code{n.update} controls number of discarded samples in warm-up stage.
\item \code{n.iter} controls how many samples are saved.
\item \code{n.thin} controls thinning, where only every \eqn{n^th} sample is kept.
\item \code{n.chains} controls number of chains.
}
}

