% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-likelihood.R
\name{mixexpert-log-lik}
\alias{mixexpert-log-lik}
\alias{extract_pointwise_log_lik}
\alias{extract_log_lik}
\title{Extract Pointwise Log Likelihood}
\usage{
extract_pointwise_log_lik(object, data)

extract_log_lik(object, data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{mixexpert}}.}

\item{data}{data frame. Used in calculation;
\code{object$data} yields in-sample result.}
}
\description{
Offline computation of log likelihood for each response variable.
}
\seealso{
The \code{loo} package for using these quantities in model comparison.
}
