% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-autoplot.R
\name{mixexpert-plot}
\alias{mixexpert-plot}
\alias{autoplot.mixexpert}
\title{MoE Patchwork Objects}
\usage{
\method{autoplot}{mixexpert}(object, varname, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link{mixexpert}}.}

\item{varname}{character. Single variable name such as \code{regr} or \code{prec}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Creates a \code{\link[patchwork:patchwork-package]{patchwork}} of \code{ggplot2} objects
with partitioned data, requires the user to add \link[=get_layer]{layers}.
}
