% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-array.R
\name{mixexpert_array}
\alias{mixexpert_array}
\alias{mixexpert_array.mcarray}
\alias{print.mixexpert_array}
\alias{tidy.mixexpert_array}
\title{MCMC Array Type}
\usage{
mixexpert_array(x)

\method{mixexpert_array}{mcarray}(x)

\method{print}{mixexpert_array}(x, fun = mean, ...)

\method{tidy}{mixexpert_array}(x, ...)
}
\arguments{
\item{x}{object. To be coerced to different format.}
}
\description{
MCMC array structure, \code{c(n_iters, n_chains, dim(.)[1], dim(.)[2], ...)},
used in this package
}
\note{
Considered and decided against packages \code{posterior}, \code{mcmcr}, \code{rjags}.
There seems to be no standardised 'best' method to store MCMC output.
}
