% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-formula.R
\name{parse_mixexpert_formula}
\alias{parse_mixexpert_formula}
\alias{mixexpert-formula}
\title{Parse MoE Formula}
\usage{
parse_mixexpert_formula(object)
}
\arguments{
\item{object}{formula. User supplied model description.}
}
\description{
Internal method to allow multiple response and multiple RHS within
a \code{\link[stats:formula]{formula}}.
}
\section{Extended Formula}{

For more information on extending \code{formula}, see the \code{Formula} package but
be warned about \url{https://github.com/rstudio/rstudio/issues/12409}.

We allow the user to jointly model multiple response variables using \code{+}:
\itemize{
\item \code{y01 + y02 ~ x01 + x02}
\item \code{y01 + y02 ~ .}
}

The RHS can also be split into two parts using \code{|} where the former is used
for regression and the latter is used for weighting.
A formula with only one part asserts a single design matrix for both.
\itemize{
\item \code{y01 ~ x01 + x02 | x03}
\item \code{y01 + y02 ~ . | x01}
}
}

\keyword{internal}
