% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-array.R
\name{pmap_bmoe_array}
\alias{pmap_bmoe_array}
\title{Apply function to each MCMC Draw}
\usage{
pmap_bmoe_array(.l, .f, ..., varname = NULL)
}
\arguments{
\item{.l}{list of \code{bmoe_array}. Only MCMC inputs should be placed here.}

\item{.f}{function. To be applied over each draws.}

\item{...}{Deterministic arguments. Same value are passed to \code{.f} each time.}

\item{varname}{character. Variable name of the stored quantity.}
}
\value{
A newly generated MCMC quantity, stored as a \code{bmoe_array}.
}
\description{
Maps a function over each MCMC iteration and chain of a \code{\link{bmoe_array}}.
}
