% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{printer_tabset}
\alias{printer_tabset}
\title{Print Tabsets (via \code{\link[knitr:knit_print]{knitr::knit_print}})}
\usage{
printer_tabset(x, options, ...)
}
\arguments{
\item{x}{list. Each element is passed to \code{knit_print} within tabs.}

\item{options, ...}{unused arguments required for \code{knit_print}.}
}
\description{
Converts a list to a Quarto / RMD tabset using the names as tab headings.
}
\references{
\url{https://github.com/nclJoshCowley/jcutils/blob/master/R/knitr-printers.R}
}
