% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

