% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_bmoe_fit}
\alias{render_bmoe_fit}
\title{Render Analysis Report}
\usage{
render_bmoe_fit(object, outfile)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{outfile}{character. Name of output file(s); file extensions dropped.}
}
\value{
Silently return the input \code{object} for use with pipes.
}
\description{
Render Quarto document and produce output files \code{.Rds} and \code{.html}.
}
