% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R, R/simulate.R
\name{example_simulate_bmoe}
\alias{example_simulate_bmoe}
\alias{example_label_switching_bmoe}
\alias{simulate_bmoe}
\title{Simulate \strong{Mixture of Experts Linear Regression}}
\usage{
example_simulate_bmoe(..., multiple_y = FALSE)

example_label_switching_bmoe()

simulate_bmoe(n_s, regr, wt, prec, n_loo, q_cens = NULL)
}
\arguments{
\item{...}{Passed to \code{simulate_bmoe}.}

\item{multiple_y}{logical. When \code{TRUE}, example defaults to multiple \code{y}.}

\item{n_s}{integer. Total number of simulated observations.}

\item{regr}{array. Regression coefficients.}

\item{wt}{matrix. Weighting coefficients.}

\item{prec}{numeric. Precision parameter per component.}

\item{n_loo}{integer. Number of observations to be moved into a test set.}

\item{q_cens}{numeric [0, 1]. Optional artificial left-censoring level.}
}
\value{
List with class \code{bmoe_sim}.
}
\description{
Simulates from a VISP model and splits into training and test splits.
}
\details{
\itemize{
\item Since \code{dim(regr) = c(n_x + 1, n_y, n_k)}, the dimension of that input
also asserts \code{n_x}, \code{n_y} and \code{n_k}.
\item Component membership is simulated per observation, not per response
variable. Hence, \code{dim(wt) = c(n_x + 1, n_k)}.
\item The simulated \code{x} matrix is used for both regression and weighting.
\item Only \code{n_s - n_loo} of the total observations will be kept in \verb{$data}.
}
}
\section{Functions}{
\itemize{
\item \code{example_label_switching_bmoe()}: Reproducible example of label switching example.

}}
