% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_mixexpert}
\alias{simulate_mixexpert}
\alias{example_simulate_mixexpert}
\title{Simulate \strong{Mixture of Experts Linear Regression}}
\usage{
simulate_mixexpert(n_s, regr, wt, prec, n_loo, q_cens = NULL)

example_simulate_mixexpert(..., multiple_y = FALSE)
}
\arguments{
\item{n_s}{integer. Number of observations.}

\item{regr}{array. Regression coefficients.}

\item{wt}{matrix. Weighting coefficients.}

\item{prec}{numeric. Precision parameter per component.}

\item{n_loo}{integer. Number of observations to be split into a test set.}

\item{q_cens}{numeric [0, 1]. Optional artificial left-censoring level.}

\item{...}{Passed to \code{simulate_mixexpert}.}

\item{multiple_y}{logical. When \code{TRUE}, example defaults to multiple \code{y}.}
}
\value{
List with class \code{mixexpert_sim}.
}
\description{
Simulates from a VISP model and splits into training and test splits.
}
\details{
\itemize{
\item Since \code{dim(regr) = c(n_x + 1, n_y, n_k)}, the dimension of that input
also asserts \code{n_x}, \code{n_y} and \code{n_k}.
\item Component membership is simulated per observation, not per response
variable. Hence, \code{dim(wt) = c(n_x + 1, n_k)}.
\item The simulated \code{x} matrix is used for both regression and weighting.
}
}
