% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_multilogit}
\alias{simulate_multilogit}
\title{Multinomial Logit Simulation}
\usage{
simulate_multilogit(x_wt, wt)
}
\arguments{
\item{x_wt}{matrix (\code{n} by \code{p_w}). Design matrix to be used in weighting.}

\item{wt}{matrix (\code{p_w} by \code{k}). Weighting regression matrix.}
}
\description{
Multinomial Logit Simulation
}
