% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_x}
\alias{simulate_x}
\title{Simulate Design Matrix}
\usage{
simulate_x(n_s, n_x, .scale = TRUE)
}
\arguments{
\item{n_s}{integer. Number of observations (rows).}

\item{n_x}{integer. Number of explanatory variables (columns).}

\item{.scale}{logical. When \code{TRUE}, output is scaled before returned.}
}
\description{
A design matrix of chosen width is simulated for use in studies.
}
\keyword{internal}
