% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{slice_by_component_values}
\alias{slice_by_component_values}
\title{Slice by Component Values}
\usage{
slice_by_component_values(x, z)
}
\description{
Delegates 'per component' slicing to get

\if{html}{\out{<div class="sourceCode">}}\preformatted{for (i in seq_along(z)) x[i, z[i]]
}\if{html}{\out{</div>}}

unlike \code{x[i, z[i]]} which returns a matrix.
}
\keyword{internal}
