% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax.R
\name{softmax}
\alias{softmax}
\alias{softmax.default}
\alias{softmax.numeric}
\alias{softmax.matrix}
\title{\code{Softmax} Function}
\usage{
softmax(x, ...)

\method{softmax}{default}(x, ...)

\method{softmax}{numeric}(x, ...)

\method{softmax}{matrix}(x, margin = 1, ...)
}
\arguments{
\item{x}{objects of vector or matrix.}

\item{...}{These dots are for future extensions and must be empty.}

\item{margin}{integer. Dimension to apply the function over,
defaults to 1 (rows).}
}
\description{
Smoothed version of "argmax"; also known as the inverse multinomial logit.
}
