% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-formula.R
\name{split_call}
\alias{split_call}
\title{Splits Call into List}
\usage{
split_call(x, sep = "|")
}
\description{
Splits Call into List
}
\seealso{
Code sourced from function body of \code{Formula:::split_formula}.
}
\keyword{internal}
