% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sweep_ref_vals}
\alias{sweep_ref_vals}
\title{Sweep Out Reference Values}
\usage{
sweep_ref_vals(x, ref = 1)
}
\arguments{
\item{x}{matrix. Matrix to be translated.}

\item{ref}{integer. Column index to be used as reference.}
}
\value{
Matrix with elements defined as \code{out[i, j] = x[i, j] - x[i, ref]}.
}
\description{
Translates matrix to satisfy corner constraint, say \code{x[, ref] = 0}.
}
