% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sweep_wt}
\alias{sweep_wt}
\title{Modify Weighting Matrix Reference Column}
\usage{
sweep_wt(wt, ref = 1)
}
\arguments{
\item{wt}{matrix. Weighting matrix.}

\item{ref}{integer. Desired reference column to be set to 0.}
}
\description{
Modify Weighting Matrix Reference Column
}
\keyword{internal}
