% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-yx.R
\name{transpose_Surv_list}
\alias{transpose_Surv_list}
\title{Splits a list of \code{Surv} objects into JAGS data}
\usage{
transpose_Surv_list(y)
}
\arguments{
\item{y}{list. Elements not inheriting \code{Surv} assumed to be uncensored data.}
}
\value{
List with \code{y} matrix and \code{is_nd} matrix.
}
\description{
Convert list of \code{Surv} objects to two matrices of identical dimension.
}
\keyword{internal}
