% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-array.R
\name{unite_and_extract}
\alias{unite_and_extract}
\title{Unite Columns into String}
\usage{
unite_and_extract(data, sep = ",")
}
\arguments{
\item{data}{data frame. Contains columns to be joined.}

\item{sep}{character. Passed to \code{paste}.}
}
\value{
character vector, \code{paste(df[, 1], ..., df[, N], sep = sep)}.
}
\description{
Unite Columns into String
}
\seealso{
\code{tidyr::unite()}.
}
\keyword{internal}
