% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{bmoe-calculate}
\alias{bmoe-calculate}
\alias{calculate_component_samples}
\title{Generated Quantities from Model Fit}
\usage{
calculate_component_samples(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}
}
\description{
Offline calculations can
}
