% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-log-lik.R
\name{bmoe-log-lik}
\alias{bmoe-log-lik}
\alias{extract_log_lik}
\alias{extract_pointwise_log_lik}
\alias{extract_componentwise_pointwise_log_lik}
\title{Extract Log Likelihood from Model Fit}
\usage{
extract_log_lik(object, new_data)

extract_pointwise_log_lik(object, new_data)

extract_componentwise_pointwise_log_lik(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}
}
\value{
\itemize{
\item \code{extract_log_lik()}
\itemize{
\item Scalar log likelihood for each MCMC iteration.
\item \code{dim(.) = c(n_iters, n_chains, 1)}.
}
}

\itemize{
\item \code{extract_pointwise_log_lik()}
\itemize{
\item Pointwise likelihood is sum of each component contribution,
weighted by the allocation probabilities.
\item \code{dim(.) = c(n_iters, n_chains, n_s_new, n_y, n_k)}.
}
}

\itemize{
\item \code{extract_componentwise_pointwise_log_lik()}
\itemize{
\item Evaluated at each component (expert) and each observation in \code{y}.
\item \code{dim(.) = c(n_iters, n_chains, n_s_new, n_y, n_k)}.
}
}
}
\description{
Offline calculations of log likelihood(s) from a \code{bmoe} model fit.
}
\seealso{
The \code{loo} package for using these quantities in model comparison.
}
