% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-likelihood.R
\name{bmoe-log-lik}
\alias{bmoe-log-lik}
\alias{calculate_pointwise_log_lik}
\alias{calculate_log_lik}
\title{Extract Pointwise Log Likelihood}
\usage{
calculate_pointwise_log_lik(object, new_data)

calculate_log_lik(object, new_data)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{new_data}{data frame.
Similar structure to original \code{data} argument.

Setting \code{new_data = NULL} signals for the observed data to be used.}
}
\description{
Offline computation of log likelihood for each response variable.
}
\seealso{
The \code{loo} package for using these quantities in model comparison.
}
