% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-package.R
\docType{package}
\name{bmoe-package}
\alias{bmoe-package}
\title{Bayesian Mixture of Experts Package}
\description{
Bayesian Mixture of Experts (MoE) applied to possibly left-censored
linear regression.
}
\section{Model Specification}{

TODO
}

