% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmoe-array.R
\name{bmoe_array}
\alias{bmoe_array}
\alias{bmoe_array.array}
\alias{bmoe_array.mcarray}
\alias{print.bmoe_array}
\alias{tidy.bmoe_array}
\title{MCMC Array Type}
\usage{
bmoe_array(x, ...)

\method{bmoe_array}{array}(x, ..., varname)

\method{bmoe_array}{mcarray}(x, ...)

\method{print}{bmoe_array}(x, fun = mean, ...)

\method{tidy}{bmoe_array}(x, ..., .dimnames = NULL)
}
\arguments{
\item{x}{object. To be coerced to different format or name used in generic.}

\item{...}{Passed on for S3 methods; ignored in \code{bmoe_array}.}

\item{varname}{character. Variable name of the stored quantity.}

\item{fun}{function. Function to summarise when printing.}

\item{.dimnames}{character.
Default behaviour only adds \verb{.term <chr>}; supplying dimension names
adds a column of indices for each dimension.}
}
\description{
MCMC array structure, \code{c(n_iters, n_chains, dim(.)[1], dim(.)[2], ...)},
used in this package.
}
\note{
For general MCMC array use, consider packages \code{posterior} and \code{mcmcr}
or the original object types in \code{rjags} and \code{coda}.
}
