% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-log-lik.R
\name{d_lcens_norm}
\alias{d_lcens_norm}
\title{Likelihood of Left-Censored Normally Distributed Vector}
\usage{
d_lcens_norm(x, mean, sd, log = FALSE)
}
\arguments{
\item{x}{object. Expected to be left-censored \code{Surv}.}

\item{mean, sd, log}{Passed to \code{\link[stats:Normal]{stats::dnorm}} and \code{\link[stats:Normal]{stats::pnorm}}.}
}
\description{
Likelihood of Left-Censored Normally Distributed Vector
}
\note{
Passing non-\code{Surv} objects lead to \code{stats::dnorm(...)}, not an error.
}
\keyword{internal}
