% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_draws}
\alias{extract_draws}
\alias{extract_draws.bmoe_fit}
\alias{extract_draws.bmoe_simstudy}
\title{Extract Posterior Samples from Bayesian Mixture of Experts}
\usage{
extract_draws(object, varname, ...)

\method{extract_draws}{bmoe_fit}(object, varname, ...)

\method{extract_draws}{bmoe_simstudy}(object, varname, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link{bmoe_fit}}.}

\item{varname}{character. Variable name defined in the model.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Extract Posterior Samples from Bayesian Mixture of Experts
}
