% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexpert-autoplot.R
\name{mixexpert-plot}
\alias{mixexpert-plot}
\alias{autoplot.mixexpert}
\title{MoE Patchwork Objects}
\usage{
\method{autoplot}{mixexpert}(object, varname, type = "none", data, ...)
}
\arguments{
\item{object}{Object that inherits from \code{\link{mixexpert}}.}

\item{varname}{character. Single variable name including \code{regr}, \code{wt},
\code{prec} and \code{log_lik}.}

\item{type}{choice. Name of the supported plot type including
\itemize{
\item \code{"none"}, plots are to be empty until the user adds layers.
\item \code{"density"}, useful distribution visualisation.
\item \code{"trace"}, useful for verifying chain convergence.
\item \code{"acf"}, autocorrelation of the MCMC samples at discrete lags.
}}

\item{data}{Passed to \link{extract_log_lik}, ignored otherwise.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Creates a \code{\link[patchwork:patchwork-package]{patchwork}} of \code{ggplot2} objects
with partitioned data, requires the user to add \link[=get_mcmc_layer]{layers}.
}
